Fixing a mix of issues

+++ kernel/src/devices/virtio/block.h
@@ -109,7 +109,7 @@ typedef struct {
 Error initVirtIOBlockDevice(int id, volatile VirtIODeviceLayout* base, VirtIODevice** output);
 
 void virtIOBlockDeviceOperation(
-    VirtIOBlockDevice* device, VirtPtr buffer, uint32_t offset, uint32_t size, bool write,
+    VirtIOBlockDevice* device, VirtPtr buffer, size_t offset, size_t size, bool write,
     VirtIOBlockCallback callback, void* udata
 );
